requirements.txt
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.3.4
numpy>=1.19.5
pillow>=8.2.0
pyyaml>=5.4.1
requests>=2.25.1
tqdm>=4.61.0
gdown>=4.4.0 